+++ src/OptionToStringGenerator/Diagnostics.cs
@@ -13,7 +13,9 @@ public static class DiagnosticTemplates
         SEEK005,
         SEEK006,
         SEEK007,
-        SEEK008
+        SEEK008,
+        SEEK009,
+        SEEK010
     }
     static List<DiagnosticDescriptor> _diagnostics = new() {
         new DiagnosticDescriptor(
@@ -68,7 +70,7 @@ public static class DiagnosticTemplates
                 category: "Usage",
                 defaultSeverity: DiagnosticSeverity.Error,
                 isEnabledByDefault: true,
-                helpLinkUri: "https://github.com/Seekatar/OptionToStringGenerator/wiki/Error-Messages#seek006-member-not-found-on-class"
+                helpLinkUri: "https://github.com/Seekatar/OptionToStringGenerator/wiki/Error-Messages#seek006-member-not-found-in-class"
                 ),
         new DiagnosticDescriptor(
                 id: Ids.SEEK007.ToString(),
@@ -87,6 +89,24 @@ public static class DiagnosticTemplates
                 defaultSeverity: DiagnosticSeverity.Error,
                 isEnabledByDefault: true,
                 helpLinkUri: "https://github.com/Seekatar/OptionToStringGenerator/wiki/Error-Messages#seek008-invalid-type-for-property"
+                ),
+        new DiagnosticDescriptor(
+                id: Ids.SEEK009.ToString(),
+                title: "Missing method for provider",
+                messageFormat: "Could not find the method '{0}' on type {1}",
+                category: "Usage",
+                defaultSeverity: DiagnosticSeverity.Error,
+                isEnabledByDefault: true,
+                helpLinkUri: "https://github.com/Seekatar/OptionToStringGenerator/wiki/Error-Messages#seek009-missing-method-for-provider"
+                ),
+        new DiagnosticDescriptor(
+                id: Ids.SEEK010.ToString(),
+                title: "Invalid provider signature",
+                messageFormat: "The signature of '{0}.{1}' should be static string? {1}({2})",
+                category: "Usage",
+                defaultSeverity: DiagnosticSeverity.Error,
+                isEnabledByDefault: true,
+                helpLinkUri: "https://github.com/Seekatar/OptionToStringGenerator/wiki/Error-Messages#seek010-invalid-provider-signature"
                 )
     };